added working api and some unit testing

+++ migrations/1673163793566_create-parking-table.js
@@ -17,7 +17,7 @@ exports.up = (pgm) => {
       notNull: true,
     },
     price: {
-      type: 'BIGINT',
+      type: 'INT',
       notNull: true,
     },
   });
